creditatlas-lit/apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.11.7
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.35.96
celery==5.4.0
redis==5.2.1
rapidfuzz==3.10.1
python-dateutil==2.9.0.post0
email-validator==2.2.0
